requirements.txt
django
mysqlclient
djangorestframework
PyJWT
django-cors-headers
djangorestframework-simplejwt